Blocks can be read in from a file and created during runtime
Not the most optimal way of  storing and making levels, but makes it easier to test while making game mechanics

+++ possesion test/src/States/PlayGround.h
@@ -18,9 +18,9 @@ public:
 
 	~PlayGround() 
 	{
-		int size = m_boxes.size();
+		size_t size = m_boxes.size();
 
-		for (int i = 0; i < size; i++)
+		for (size_t i = 0; i < size; i++)
 		{
 			delete m_boxes[i];
 		}
